Move config to universal
