+++ src/client/config/app.ts
@@ -1,3 +1,7 @@
+// Legacy export, these used to live in this file
+export { Colors } from '../../universal/config/colors';
+export { ExternalUrls } from './external-urls';
+
 // NOTE: Keep up-to-date with _breakpoints.scss
 export const Breakpoints = {
   tablet: 1024,
@@ -8,9 +12,5 @@ export const Breakpoints = {
 
 export const PLACEHOLDER_IMAGE_URL =
   'data:image/gif;base64,R0lGODlhAQABAIAAAMLCwgAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==';
-
-// Legacy export, these used to live in this file
-export { Colors } from '../../universal/config/feature-toggles';
-export { ExternalUrls } from './external-urls';
 
 export const MAX_TABLE_ROWS_ON_THEMA_PAGINA = 3;
